native://: the board layer's drive path on this host in real time, for SIL/HIL
board_pwm.c, board_sync.c, board_adc.c and board_drive.c built unchanged over
board/native's TIM1 (preload, RCR, UEV, rc_w0 SR), ADC1-3 and AFE (Renode's
models in C), the world's plant between the timer's edges. Paced 1.000, 50 000
triples a wall second, no overrun; 1 virtual s in 0.048 wall s flat out.
Validation stays on Renode. test_structure: the target builds nothing of
board/fake, board/native, board/emu; no firmware source asks where it runs.

The A1335 and BNO085 pause for the observer's AFE_ON borrow only when it holds
the rail alone (Board_PowerAlone): with the host holding it they dropped 0.5 s
every 30 s. The ring test stamps a read mid-way: a stall under load swapped
neck and head.

+++ board/src/board_pwm.c
@@ -594,7 +594,7 @@ static uint32_t dts_ps(void)
   /* Picoseconds per DTG count, so the ns arithmetic stays integer. */
   const uint32_t hz = Board_SysClkHz() / 2UL;   /* TIM1 kernel, 237.5 MHz */
 
-  return (hz != 0UL) ? (PS_PER_S / hz) : 0UL;
+  return (hz != 0UL) ? (uint32_t)(PS_PER_S / hz) : 0UL;
 }
 
 uint8_t Board_PwmDeadTimeFloor(void)
@@ -691,7 +691,7 @@ const char *Board_PwmSetDeadTime(uint32_t ns)
   /* Rounded up, like the floor above and for the same reason: a dead time
      that rounded down is under what was asked for, and the direction that is
      wrong is the one that shortens it. */
-  uint32_t counts = (((uint64_t)ns * PS_PER_NS) + ps - 1ULL) / ps;
+  uint32_t counts = (uint32_t)((((uint64_t)ns * PS_PER_NS) + ps - 1ULL) / ps);
   const uint8_t floor_counts = Board_PwmDeadTimeFloor();
 
   if (counts < floor_counts)
